refactor(#1983 Finding-3): single-source OP_KIND_MODEL_MAP in models.py + derive ControlIROp union; enforce mcp_drop_server
The ControlIROp union (models.py) was hand-listed SEPARATELY from the authoritative
OP_KIND_MODEL_MAP (op_runtime/registry.py) — a dual-source. mcp_drop_server was
registered + documented (control-ir.md) + modeled + handled but ABSENT from both
the union and the map → a phase emitting it failed ActOutput validation
(advertise-but-don't-enforce; the worst direction of the control-ir.md ↔ map sync
invariant).

Root fix (lead-approved D1 = relocate, clean-break): the map's LOCATION was why the
union couldn't derive from it — registry.py imports the IROp model classes, so
models.py importing registry's map would cycle. So:
- RELOCATE OP_KIND_MODEL_MAP + ALL_OP_KINDS into schemas/models.py (co-located with
  the IROp classes + the union). registry.py keeps OP_PURITY and re-imports
  ALL_OP_KINDS from models for ALL_TOOL_NAMES (intentional convenience, documented;
  not a migration shim).
- DERIVE the ControlIROp union from the map: Union[(FileIROp, *OP_KIND_MODEL_MAP
  .values())] — completeness-by-construction (a kind in the map is in the union, no
  dual-source). FileIROp (coarse "file") is the only explicit non-map member
  (internal-use). A TYPE_CHECKING static mirror keeps mypy precise (runtime derives).
- ADD mcp_drop_server: map + OP_PURITY(side_effect) + _OP_KIND_ALIASES + the
  phase-dispatch wired set. The union + ALL_OP_KINDS follow by construction.
- REPOINT all ~9 importers (control_ir_executor + 8 tests) registry → models
  (clean-break, no shim; OP_PURITY/OpPurity stay registry). CLAUDE.md sync-rule ref
  updated (registry.py → models.py).

Verified: imports clean (no cycle); the derived union validates mcp_drop_server +
is complete-by-construction; 273 tests pass across the changed-module importers;
ruff clean. Finding-4 (op_catalog sourced from the map) + the completeness-invariant
& falsifying tests follow in the next commit.
